stage1: Move common code to init
It is used by init only, so let's keep the code in init
directory. This package is meant to be used by both kvm and
systemd-nspawn code paths in init.
